Added Grass Struct with Wild Pokémon List and Encounter Rate
Created the Grass struct to represent a patch of grass containing wild Pokémon.
- Added a vector to store WildPokemon structs, representing the Pokémon that can hide in the grass.
- Introduced an encounter rate (integer) to specify how often Pokémon appear in the grass.
- Added an environment description (string) to specify where the grass is located (e.g., "Forest", "Cave").
