+++ .gitignore
@@ -30,3 +30,5 @@
 *.exe
 *.out
 *.app
+
+*.vs
