+++ Pokemon/Game.hpp
@@ -1,0 +1,10 @@
+//Game.hpp
+#include "Grass.hpp"
+class Player;
+
+class Game
+{
+public:
+	Game();
+	void gameLoop(Player& player);
+};
